requirements.txt
beautifulsoup4==4.12.3
biopython==1.84
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
html5lib==1.1
kiwisolver==1.4.7
Levenshtein==0.26.0
lxml==5.3.0
matplotlib==3.9.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-Levenshtein==0.26.0
pytz==2024.2
RapidFuzz==3.10.0
seaborn==0.13.2
six==1.16.0
soupsieve==2.6
tzdata==2024.2
webencodings==0.5.1
